# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "caduceus API",
    "description": "REST API of the caduceus collaborative Typst editor. Protected routes authenticate via a JWT, sent either as an HttpOnly `token` cookie or an `Authorization: Bearer` header.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/login": {
      "post": {
        "tags": [
          "user"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Logged in; also sets the `token` cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess_AuthPayload"
                }
              }
            }
          },
          "401": {
            "description": "Wrong password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/register": {
      "post": {
        "tags": [
          "user"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User registered; also sets the `token` cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess_AuthPayload"
                }
              }
            }
          },
          "400": {
            "description": "Password too long",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "409": {
            "description": "Username already taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          }
        }
      }
    },
    "/api/team": {
      "post": {
        "tags": [
          "team"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Team created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess_TeamPayload"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "404": {
            "description": "Creator not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          }
        }
      }
    },
    "/api/team/projects": {
      "get": {
        "tags": [
          "team"
        ],
        "operationId": "projects",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Team id (24-char hex ObjectId)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Projects owned by the team",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess_Vec_ProjectPayload"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "404": {
            "description": "Team not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/me": {
      "get": {
        "tags": [
          "user"
        ],
        "operationId": "me",
        "responses": {
          "200": {
            "description": "The current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess_UserPayload"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/projects": {
      "get": {
        "tags": [
          "user"
        ],
        "operationId": "projects",
        "responses": {
          "200": {
            "description": "Projects the current user can access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess_Vec_ProjectPayload"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/teams": {
      "get": {
        "tags": [
          "user"
        ],
        "operationId": "teams",
        "responses": {
          "200": {
            "description": "Teams the current user belongs to",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess_Vec_TeamPayload"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiMessage": {
        "type": "object",
        "description": "Doc-only mirror of [`crate::models::response::ApiResponse`] for bodies\nwithout a payload: every error response, plus payload-less successes such\nas logout. `#[serde(skip_serializing_if)]` removes the `payload` key\nentirely in this case, so the wire shape is just `{ \"message\": ... }`.",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "ApiSuccess_AuthPayload": {
        "type": "object",
        "description": "Doc-only mirror of [`crate::models::response::ApiResponse`] for responses\nthat always carry a payload. `ApiResponse.payload` is an `Option` only so\none struct can serve both cases; on the wire a success body is always\n`{ \"message\": ..., \"payload\": ... }`, which is what this schema states\n(a `ToSchema` derive on the real struct would wrongly mark `payload`\noptional and nullable). A test below pins this equivalence.",
        "required": [
          "message",
          "payload"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "payload": {
            "type": "object",
            "required": [
              "user",
              "token"
            ],
            "properties": {
              "token": {
                "type": "string"
              },
              "user": {
                "$ref": "#/components/schemas/UserPayload"
              }
            }
          }
        }
      },
...
      "ApiSuccess_TeamPayload": {
        "type": "object",
        "description": "Doc-only mirror of [`crate::models::response::ApiResponse`] for responses\nthat always carry a payload. `ApiResponse.payload` is an `Option` only so\none struct can serve both cases; on the wire a success body is always\n`{ \"message\": ..., \"payload\": ... }`, which is what this schema states\n(a `ToSchema` derive on the real struct would wrongly mark `payload`\noptional and nullable). A test below pins this equivalence.",
        "required": [
          "message",
          "payload"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "payload": {
            "type": "object",
            "required": [
              "id",
              "name",
              "avatar_uri",
              "creator_id",
              "member_ids",
              "created_at",
              "updated_at"
            ],
            "properties": {
              "avatar_uri": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "creator_id": {
                "type": "string"
              },
              "id": {
                "type": "string"
              },
              "member_ids": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "name": {
                "type": "string"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        }
      },
...
      "ApiSuccess_UserPayload": {
        "type": "object",
        "description": "Doc-only mirror of [`crate::models::response::ApiResponse`] for responses\nthat always carry a payload. `ApiResponse.payload` is an `Option` only so\none struct can serve both cases; on the wire a success body is always\n`{ \"message\": ..., \"payload\": ... }`, which is what this schema states\n(a `ToSchema` derive on the real struct would wrongly mark `payload`\noptional and nullable). A test below pins this equivalence.",
        "required": [
          "message",
          "payload"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "payload": {
            "type": "object",
            "required": [
              "id",
              "username",
              "nickname",
              "avatar_uri",
              "created_at",
              "updated_at"
            ],
            "properties": {
              "avatar_uri": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "id": {
                "type": "string"
              },
              "nickname": {
                "type": "string"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time"
              },
              "username": {
                "type": "string"
              }
            }
          }
        }
      },
      "ApiSuccess_Vec_ProjectPayload": {
        "type": "object",
        "description": "Doc-only mirror of [`crate::models::response::ApiResponse`] for responses\nthat always carry a payload. `ApiResponse.payload` is an `Option` only so\none struct can serve both cases; on the wire a success body is always\n`{ \"message\": ..., \"payload\": ... }`, which is what this schema states\n(a `ToSchema` derive on the real struct would wrongly mark `payload`\noptional and nullable). A test below pins this equivalence.",
        "required": [
          "message",
          "payload"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "payload": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "name",
                "owner_id",
                "owner_type",
                "creator_id",
                "files",
                "created_at",
                "updated_at",
                "entry",
                "pinned_version"
              ],
              "properties": {
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "creator_id": {
                  "type": "string"
                },
                "entry": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "files": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectFilePayload"
                  }
                },
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "owner_id": {
                  "type": "string"
                },
                "owner_type": {
                  "$ref": "#/components/schemas/OwnerType"
                },
                "pinned_version": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "updated_at": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          }
        }
      },
      "ApiSuccess_Vec_TeamPayload": {
        "type": "object",
        "description": "Doc-only mirror of [`crate::models::response::ApiResponse`] for responses\nthat always carry a payload. `ApiResponse.payload` is an `Option` only so\none struct can serve both cases; on the wire a success body is always\n`{ \"message\": ..., \"payload\": ... }`, which is what this schema states\n(a `ToSchema` derive on the real struct would wrongly mark `payload`\noptional and nullable). A test below pins this equivalence.",
        "required": [
          "message",
          "payload"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "payload": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "name",
                "avatar_uri",
                "creator_id",
                "member_ids",
                "created_at",
                "updated_at"
              ],
              "properties": {
                "avatar_uri": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "creator_id": {
                  "type": "string"
                },
                "id": {
                  "type": "string"
                },
                "member_ids": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "name": {
                  "type": "string"
                },
                "updated_at": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          }
        }
      },
...
      "CreateTeamRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "FileKind": {
        "type": "string",
        "enum": [
          "text",
          "binary"
        ]
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "OwnerType": {
        "type": "string",
        "enum": [
          "user",
          "team"
        ]
      },
...
      "ProjectFilePayload": {
        "type": "object",
        "description": "File metadata for listings. Deliberately does NOT inline `content`: the\ndirectory/tree view only needs path + kind, and the body is fetched\nper-file when a tab is opened.",
        "required": [
          "id",
          "path",
          "kind",
          "size",
          "version",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/FileKind"
          },
          "path": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "UserPayload": {
        "type": "object",
        "required": [
          "id",
          "username",
          "nickname",
          "avatar_uri",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "avatar_uri": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "username": {
            "type": "string"
          }
        }
      }
    }
  },
...
}